Added productName detection in package.json
Electron apps often have a productName field that is used by the builder
which can be different from the name filed. This will cause creation of
multiple config folders, one for the app's Electron stuff, and one for
your configs. Prefer productName over name, fall back to name if
productName not defined.